+++ app/build.gradle
@@ -51,6 +51,7 @@ dependencies {
     implementation 'androidx.navigation:navigation-ui-ktx:2.3.0'
     implementation 'androidx.preference:preference:1.1.1'
     implementation "com.google.dagger:hilt-android:2.28-alpha"
+    implementation 'androidx.hilt:hilt-lifecycle-viewmodel:1.0.0-alpha02'
     kapt "com.google.dagger:hilt-android-compiler:2.28-alpha"
     kapt 'androidx.room:room-compiler:2.2.5'
     implementation "androidx.room:room-runtime:2.2.5"
