+++ app/src/main/java/com/project/segunfrancis/superherocollection/di/DatabaseModule.kt
@@ -1,6 +1,7 @@
 package com.project.segunfrancis.superherocollection.di
 
 import android.content.Context
+import androidx.room.Room
 import com.project.segunfrancis.superherocollection.framework.local.SuperHeroDao
 import com.project.segunfrancis.superherocollection.framework.local.SuperHeroRoomDatabase
 import dagger.Module
@@ -21,7 +22,15 @@ object DatabaseModule {
     @Provides
     @Singleton
     fun provideDatabase(@ApplicationContext context: Context): SuperHeroRoomDatabase {
-        return SuperHeroRoomDatabase.getInstance(context.applicationContext)!!
+        var superHeroRoomInstance: SuperHeroRoomDatabase
+        synchronized(SuperHeroRoomDatabase::class.java) {
+            superHeroRoomInstance = Room.databaseBuilder<SuperHeroRoomDatabase>(
+                context.applicationContext,
+                SuperHeroRoomDatabase::class.java,
+                "super_hero_database"
+            ).build()
+        }
+        return superHeroRoomInstance
     }
 
     @Provides
